feat: model object built with training capability
The model object is now in an executable state and can be
run from the command line. The different command line arguments
are also integrated so users can modify some of the model's
training parameters.

A method for saving a checkpoint is also included; however, it
still needs to be tested.

Currently, the program should work on either a CPU or GPU. This
functionality also needs to be tested under either condition to
be sure it defaults appropriately.