feat: started module for predict.py
Started building on the FCNN object to also have the ability to
work with the predict capabilities. The FCNN object code was
moved into a separate python file for inclusion with both train.py
and predict.py.

The methods for the predict related capabilities was also added
to the fcnn object. The code needs to be refactored now and made
compatible with both sets of major functionality (train & predict).