docs: updated the docstrings for all methods
Not all the docstrings for the individual methods and
functions were correct. These should all be updated to represent
the parameters and return values correctly.

There were also a number of bug fixes that needed to be dealt
with and both programs (train.py and predict.py) have tested
successfully now.